test4

+++ modules/task_3/kolesova_k_sort_betch/main.cpp
@@ -11,10 +11,10 @@
 TEST(Sort_Batcher, Vec_Setpoints) {
   int rank;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-  std::vector<int> vec = {  103, 5, 10, 21, 35, 18, 13, 29};
-  std::vector<int> res = { 5, 10, 13, 18, 21, 29, 35, 103};
+  std::vector<double> vec = {  103, 5, 10, 21, 35, 18, 13, 29};
+  std::vector<double> res = { 5, 10, 13, 18, 21, 29, 35, 103};
 
-  std::vector<int> parralel = merge_batcher(vec, vec.size());
+  std::vector<double> parralel = merge_batcher(vec, vec.size());
   if (rank == 0) {
     ASSERT_EQ(parralel, res);
   }
@@ -23,13 +23,13 @@ TEST(Sort_Batcher, Vec_Setpoints) {
 TEST(Sort_Batcher, Vec_50_Elem) {
   int rank;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-  std::vector<int> global_vec;
+  std::vector<double> global_vec;
   const int size_vector = 50;
   if (rank == 0) {
     global_vec = randVec(size_vector);
   }
 
-  std::vector<int> parralel = merge_batcher(global_vec, size_vector);
+  std::vector<double> parralel = merge_batcher(global_vec, size_vector);
   if (rank == 0) {
     global_vec = sort(global_vec);
     ASSERT_EQ(parralel, global_vec);
@@ -39,13 +39,13 @@ TEST(Sort_Batcher, Vec_50_Elem) {
 TEST(Sort_Batcher, Vec_100_Elem) {
   int rank;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-  std::vector<int> global_vec;
+  std::vector<double> global_vec;
   const int size_vector = 100;
   if (rank == 0) {
     global_vec = randVec(size_vector);
   }
 
-  std::vector<int> parralel = merge_batcher(global_vec, size_vector);
+  std::vector<double> parralel = merge_batcher(global_vec, size_vector);
   if (rank == 0) {
     global_vec = sort(global_vec);
     ASSERT_EQ(parralel, global_vec);
